+++ app/Http/Controllers/AttributeController.php
@@ -9,17 +9,20 @@ class AttributeController extends Controller
 {
     public function index()
     {
+        $this->authorize('attribute_access');
     	$attributes = Attribute::paginate(25);
     	return view('admin/attributes/attributes')->with('attributes', $attributes);
     }
 
     public function create()
     {
+        $this->authorize('attribute_create');
     	return view('admin/attributes/attribute_create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('attribute_create');
     	$request->validate(['attribute' => 'required|unique:attributes']);
     	$slug = str_replace(' ', '-', strtolower(trim($request->post('attribute'))));
     	$slug = preg_replace('/[^A-Za-z0-9\-]/', '', $slug);
@@ -33,12 +36,14 @@ class AttributeController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('attribute_update');
     	$attribute = Attribute::find($id);
     	return view('admin/attributes/attribute_edit')->with('attribute', $attribute);
     }
 
     public function update(Request $request)
     {
+        $this->authorize('attribute_update');
     	$request->validate(['attribute' => 'required|unique:attributes']);
     	$slug = str_replace(' ', '-', strtolower(trim($request->post('attribute'))));
     	$slug = preg_replace('/[^A-Za-z0-9\-]/', '', $slug);
@@ -53,6 +58,7 @@ class AttributeController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('attribute_delete');
     	Attribute::where('id', $id)->delete();
     	return redirect('admin/attributes')->withErrors('DELETED !! Attribute is successfully deleted');
     }
