+++ app/Http/Controllers/RoleController.php
@@ -9,17 +9,20 @@ class RoleController extends Controller
 {
     public function index()
     {
+        $this->authorize('role_access');
         $roles = Role::get()->all();
         return view('admin/roles/roles')->with('roles', $roles);
     }
 
     public function create($value='')
     {
+        $this->authorize('role_create');
     	return view('admin/roles/role_create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('role_create');
         $request->validate([
             'role_name'  =>  'required|unique:roles',
         ]);
@@ -29,12 +32,14 @@ class RoleController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('role_update');
     	$role = Role::find($id);
     	return view('admin/roles/role_edit')->with('role', $role);
     }
 
     public function update(Request $request)
     {
+        $this->authorize('role_update');
     	$request->validate([
     	    'role_name'  =>  'required',
     	]);
@@ -45,6 +50,7 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('role_delete');
         Role::where('id', $id)->delete();
         return redirect()->back()
                         ->withErrors(['DELETED !! Role is successfully delete']);
