+++ app/Http/Controllers/RolePermissionController.php
@@ -11,7 +11,7 @@ class RolePermissionController extends Controller
 {
     public function index($value='')
     {
-        //$this->authorize('role_permission_access');
+        $this->authorize('role_permission_access');
     	$roles = Role::with('permissions')->get()->all();
 		$permissions = Permission::with('children')
 	                                ->whereNull('parent')
@@ -24,7 +24,7 @@ class RolePermissionController extends Controller
 
     public function update(Request $request)
     {
-        //$this->authorize('role_permission_update');
+        $this->authorize('role_permission_update');
     	$data = $request->all();
 
     	Role_permission::where('role_id', $data['role_id'])->delete();
